--- frontend/src/app/Pages/image/image.component.html
@@ -1,17 +1,0 @@
-<div class="card" style="width: 25rem;">
-    <div class="card-body">
-        <form [formGroup]="formTemplate" (submit)="onSubmit(formTemplate.value)">
-            <div class="text-center">
-                <img [src]="imgSrc" width="350px" height="250px" (click)="fileUploader.click()">
-            </div>
-            <div class="form-group">
-                <label>ファイルをアップロード</label>
-                <input type="file" accept="image/png,image/jpg" class="form-control" #fileUploader formControlName="imageUrl" (change)="showPreview($event)">
-                <div class="text-danger" *ngIf="isSubmitted && formControls.imageUrl.errors?.['required']">画像を追加してください。</div>
-            </div>
-            <div class="form-group">
-                <button type="submit" class="btn btn-success float-right">Submit</button>
-            </div>
-        </form>
-    </div>
-</div>
