+++ frontend/src/app/app-routing.module.ts
@@ -3,7 +3,6 @@ import { RouterModule, Routes } from '@angular/router';
 import { AuthGuard } from './guards/auth.guard';
 import { ClothesEditComponent } from './pages/clothes-edit/clothes-edit.component';
 import { ClothesListComponent } from './pages/clothes-list/clothes-list.component';
-import { ImageComponent } from './pages/image/image.component';
 import { ClothesAddComponent } from './pages/clothes-add/clothes-add.component';
 import { LiffInitComponent } from './pages/liff-init/liff-init.component';
 
@@ -17,12 +16,6 @@ const routes: Routes = [
     path: 'clothes-list',
     canActivate: [AuthGuard],
     component: ClothesListComponent,
-  },
-  {
-    path: 'image',
-    component: ImageComponent,children:[
-      {path: 'upload', component:ImageComponent} //image/upload
-    ]
   },
   {
     path: 'clothes-edit',
